requirements.txt
# Build tools
pip>=24.0
setuptools>=69.5.1
wheel>=0.43.0

# Core dependencies
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.25.2
aiohttp==3.8.6
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal==1.4.0
APScheduler==3.9.1.post1
asgiref==3.9.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.9.1
bidict==0.23.1
blis==0.7.11
boto3==1.39.4
botocore==1.39.4
CacheControl==0.12.14
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.5
confluent-kafka==2.11.0
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
dask==2022.10.2
dj-database-url==3.0.1

# Django core
Django==5.2.4
django-allauth==65.10.0
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cloudinary-storage==0.3.0
django-countries==7.6.1
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular

# Other dependencies
dnspython==2.3.0
docopt==0.6.2
fbmessenger==6.0.0
filelock==3.18.0
fire==0.5.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.62.3
gunicorn==20.1.0
h11==0.16.0
h5py==3.14.0
httplib2==0.22.0
httptools==0.6.4
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
# REMOVED: JAX and jaxlib (not needed for Django + Rasa)
# jax==0.4.23
# jaxlib==0.4.23
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
locket==1.0.0
magic-filter==1.0.12
marisa-trie==1.2.1
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.5.3
mattermostwrapper==2.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
multidict==5.2.0
murmurhash==1.0.13
networkx==2.6.3
nltk==3.9.1

# FIXED: Compatible with TensorFlow 2.12.0 and Rasa 3.6.16
numpy==1.23.5

oauthlib==3.3.1
opt_einsum==3.4.0
packaging==20.9
pamqp==3.2.1
pandas==1.5.3
partd==1.4.2
pathlib_abc==0.1.1
pathy==0.11.0
pluggy==1.6.0
portalocker==2.10.1
preshed==3.0.10
progressbar2==4.5.0
prompt-toolkit==3.0.28
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.23.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.7
pydot==1.4.2
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.2.3
pyreadline3==3.5.4
pyrsistent==0.20.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
python-utils==3.9.1
pytz==2022.7.1
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5

# Rasa and related
rasa==3.6.16
rasa-sdk==3.6.2

redis==4.6.0
regex==2022.10.31
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.28.1
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
safetensors==0.5.3
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==1.14.0
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
slack_sdk==3.36.0
smart-open==6.4.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
sqlparse==0.4.3
srsly==2.5.1
structlog==23.3.0
structlog-sentry==2.1.0
sympy==1.14.0
tabulate==0.9.0
tarsafe==0.0.4

# TensorFlow - FIXED: Changed from just "tensorflow" to specific version
tensorflow==2.12.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0

termcolor==3.1.0
terminaltables==3.1.10
thinc==8.1.12
threadpoolctl==3.6.0
tokenizers==0.13.3
toolz==1.0.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.30.2
twilio==8.2.2
typer==0.7.0
typing-utils==0.1.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.20.1

# Add these lines to your existing requirements.txt file
SpeechRecognition>=3.10.0
pydub>=0.25.1
sounddevice>=0.4.8
celery>=5.3.0
django-redis>=5.2.0
kombu>=5.3.0
billiard>=4.1.0
vine>=5.0.0
gtts==2.3.2
Pillow==10.0.0